Moved containers too, got rid of components directory
