+++ meteor-app/imports/ui/member/member-card.js
