+++ meteor-app/imports/ui/member/member-container.js
