+++ meteor-app/imports/ui/member/member-list.js
