+++ meteor-app/imports/ui/member/member-search.js
