+++ meteor-app/imports/ui/member/member-visit-container.js
