+++ meteor-app/tests/mocha.storyspecs.opts
@@ -3,6 +3,6 @@
 --require  ./.storybook/__conf__/mochaMockConfig.js
 
 imports/ui/**/**/*.stories.js
---compilers js:babel-core/register
+--require babel-core/register
 --recursive
